+++ src/core/mouse/windows/mod.rs
@@ -1,12 +1,15 @@
 use crate::core::mouse::{MouseClick, MouseScroll};
+use crate::errors::AutoGuiError;
 use std::mem::{size_of, zeroed};
 use std::{thread, time, time::Instant};
-use winapi::shared::windef::POINT;
-use winapi::um::winuser::{
-    SendInput, SetCursorPos, INPUT, INPUT_MOUSE, MOUSEEVENTF_HWHEEL, MOUSEEVENTF_LEFTDOWN,
-    MOUSEEVENTF_LEFTUP, MOUSEEVENTF_MIDDLEDOWN, MOUSEEVENTF_MIDDLEUP, MOUSEEVENTF_RIGHTDOWN,
-    MOUSEEVENTF_RIGHTUP, MOUSEEVENTF_WHEEL,
+use windows::Win32::Foundation::POINT;
+use windows::Win32::UI::Input::KeyboardAndMouse::{
+    SendInput, INPUT, INPUT_MOUSE, MOUSEEVENTF_HWHEEL, MOUSEEVENTF_LEFTDOWN, MOUSEEVENTF_LEFTUP,
+    MOUSEEVENTF_MIDDLEDOWN, MOUSEEVENTF_MIDDLEUP, MOUSEEVENTF_RIGHTDOWN, MOUSEEVENTF_RIGHTUP,
+    MOUSEEVENTF_WHEEL,
 };
+use windows::Win32::UI::WindowsAndMessaging::{GetCursorPos, SetCursorPos};
+
 #[derive(Debug)]
 pub struct Mouse {}
 impl Mouse {
@@ -16,17 +19,16 @@ impl Mouse {
     }
 
     /// moves mouse to x, y pixel coordinate on screen
-    pub fn move_mouse_to_pos(x: i32, y: i32, moving_time: f32) {
+    pub fn move_mouse_to_pos(x: i32, y: i32, moving_time: f32) -> Result<(), AutoGuiError> {
         // if no moving time, then instant move is executed
         unsafe {
             if moving_time <= 0.0 {
-                SetCursorPos(x, y);
-                return;
+                return SetCursorPos(x, y).map_err(|err| err.into());
             }
         };
         // if moving time is included, loop is executed that moves step by step
         let start = Instant::now();
-        let start_location = Mouse::get_mouse_position();
+        let start_location = Mouse::get_mouse_position()?;
         let distance_x = x - start_location.0;
         let distance_y = y - start_location.1;
 
@@ -44,42 +46,45 @@ impl Mouse {
 
             unsafe {
                 if time_passed_percentage >= 1.0 {
-                    SetCursorPos(x, y);
-                    break;
+                    return SetCursorPos(x, y).map_err(|err| err.into());
                 } else {
-                    SetCursorPos(new_x as i32, new_y as i32);
+                    let result = SetCursorPos(new_x as i32, new_y as i32).map_err(|err| err.into());
+                    if result.is_err() {
+                        return result;
+                    }
                 }
             }
         }
     }
 
-    pub fn drag_mouse(x: i32, y: i32, moving_time: f32) {
+    pub fn drag_mouse(x: i32, y: i32, moving_time: f32) -> Result<(), AutoGuiError> {
         let (down, up) = (MOUSEEVENTF_LEFTDOWN, MOUSEEVENTF_LEFTUP);
         unsafe {
             // set up the first input event (mouse down)
             let mut input_down: INPUT = zeroed();
-            input_down.type_ = INPUT_MOUSE;
-            input_down.u.mi_mut().dwFlags = down;
-            SendInput(1, &mut input_down, size_of::<INPUT>() as i32);
+            input_down.r#type = INPUT_MOUSE;
+            input_down.Anonymous.mi.dwFlags = down;
+            SendInput(&[input_down], size_of::<INPUT>() as i32);
             // wait a bit after click down, before moving
             thread::sleep(time::Duration::from_millis(80));
-            Mouse::move_mouse_to_pos(x, y, moving_time);
+            Mouse::move_mouse_to_pos(x, y, moving_time)?;
             thread::sleep(time::Duration::from_millis(50));
             // set up the second input event (mouse up)
             let mut input_up: INPUT = zeroed();
-            input_up.type_ = INPUT_MOUSE;
-            input_up.u.mi_mut().dwFlags = up;
+            input_up.r#type = INPUT_MOUSE;
+            input_up.Anonymous.mi.dwFlags = up;
             // send the input events
-            SendInput(2, &mut input_up, size_of::<INPUT>() as i32);
+            SendInput(&[input_up], size_of::<INPUT>() as i32);
+            Ok(())
         }
     }
 
     /// returns x, y pixel coordinate of mouse position
-    pub fn get_mouse_position() -> (i32, i32) {
+    pub fn get_mouse_position() -> Result<(i32, i32), AutoGuiError> {
         unsafe {
             let mut point = POINT { x: 0, y: 0 };
-            winapi::um::winuser::GetCursorPos(&mut point);
-            (point.x, point.y)
+            GetCursorPos(&mut point)?;
+            Ok((point.x, point.y))
         }
     }
 
@@ -95,13 +100,13 @@ impl Mouse {
             // create an array of INPUT structures
             let mut inputs: [INPUT; 2] = [zeroed(), zeroed()];
             // set up the first input event (mouse down)
-            inputs[0].type_ = INPUT_MOUSE;
-            inputs[0].u.mi_mut().dwFlags = down;
+            inputs[0].r#type = INPUT_MOUSE;
+            inputs[0].Anonymous.mi.dwFlags = down;
             // set up the second input event (mouse up)
-            inputs[1].type_ = INPUT_MOUSE;
-            inputs[1].u.mi_mut().dwFlags = up;
+            inputs[1].r#type = INPUT_MOUSE;
+            inputs[1].Anonymous.mi.dwFlags = up;
             // send the input events
-            SendInput(2, inputs.as_mut_ptr(), size_of::<INPUT>() as i32);
+            SendInput(&inputs, size_of::<INPUT>() as i32);
         }
     }
 
@@ -116,11 +121,11 @@ impl Mouse {
             // create an array of INPUT structures
             let mut input: INPUT = zeroed();
 
-            input.type_ = INPUT_MOUSE;
-            input.u.mi_mut().dwFlags = down;
+            input.r#type = INPUT_MOUSE;
+            input.Anonymous.mi.dwFlags = down;
 
             // send the input events
-            SendInput(1, &mut input, size_of::<INPUT>() as i32);
+            SendInput(&[input], size_of::<INPUT>() as i32);
         }
     }
 
@@ -135,11 +140,11 @@ impl Mouse {
             // create an array of INPUT structures
             let mut input: INPUT = zeroed();
             // set up thefirstut vent (mous;
-            input.type_ = INPUT_MOUSE;
-            input.u.mi_mut().dwFlags = up;
+            input.r#type = INPUT_MOUSE;
+            input.Anonymous.mi.dwFlags = up;
 
             // send the input events
-            SendInput(1, &mut input, size_of::<INPUT>() as i32);
+            SendInput(&[input], size_of::<INPUT>() as i32);
         }
     }
 
@@ -155,10 +160,10 @@ impl Mouse {
         unsafe {
             let mut scroll_input: INPUT = zeroed();
 
-            scroll_input.type_ = INPUT_MOUSE;
-            scroll_input.u.mi_mut().dwFlags = wheel_direction;
-            scroll_input.u.mi_mut().mouseData = amount as u32;
-            SendInput(1, &mut scroll_input, size_of::<INPUT>() as i32);
+            scroll_input.r#type = INPUT_MOUSE;
+            scroll_input.Anonymous.mi.dwFlags = wheel_direction;
+            scroll_input.Anonymous.mi.mouseData = amount as u32;
+            SendInput(&[scroll_input], size_of::<INPUT>() as i32);
         }
     }
 }