+++ src/errors/windows/mod.rs
@@ -1,0 +1,7 @@
+use crate::errors::AutoGuiError;
+
+impl From<windows_result::Error> for AutoGuiError {
+    fn from(err: windows_result::Error) -> Self {
+        AutoGuiError::OSFailure(err.message())
+    }
+}